A0-2582: Extract e2e tests to a separate workflow (#1227)
# Description

There's an issue with the merge queue so that on[
`merge_group:`](https://github.com/Cardinal-Cryptography/aleph-node/actions/workflows/on-pull-request-commit.yml?query=event%3Amerge_group)
trigger, when an e2e (e.g. flaky) fails, then commit is merged to
`main`, whereas it should be reverted from the queue. [Example 1
](https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/5078761334)and
[Example
2.](https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/5088540897)

A common behavior of this example is that last job, `check e2e suite
test completion` is canceled when any of e2e tests fails. This is
expected though according to official docs
> If a job fails or is skipped, all jobs that need it are skipped unless
the jobs use a conditional expression that causes the job to continue
yet, such PR should not be merged to `main` by merge queue as `check e2e
suite test completion` is a required check.

This PR attempts to fix above, and `main` situation will be monitored in
case more fixes are needed. For sure it makes workflows more clear as
now A) there's no need to have a duplication list of e2e jobs B) PR
commit workflow is quite clear code now.

## Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)