A0-1898: add more timeouts in clique network (#1286)
# Description

Since by default tokio's send operations for tcp connections do not
timeout, it was possible to exploit a node and force it to create any
number of incoming connections with a single node (Replace mechanic not
working as expected) without possibility to drop previous ones - task
was unable to receive a signal from a dropped user_data channel and
finish.
NOTE: this issue could be also solved on the clique-network service
level, which would also allow to drop on-going send or even handshake
procedures - current version forces a task to wait for them to timeout
and then checks that channel. We chose this solution for its simplicity,
since implementing it on the service level would require a lot of
refactoring.
There's also os' internal buffer for tcp sockets, in my case around
15kb, so stopping `receive` is not visible on the sender's site
immediately.

## Type of change

- Bug fix (non-breaking change which fixes an issue)