A0-3053: Fix contracts/scripts/clean.sh (#1357)
# Description

Cargo-contract 3.0.0 is not 100% compatible with our pallet-contracts,
so we need to use this older version to terminate contracts and remove
code. Incidentally, I removed the cliain image from this script, so we
only use 2 images instead of 3 different ones (also, it wasn't quite
working anyway).

Should fix CI for button.

## Type of change

- Bug fix (non-breaking change which fixes an issue)