A0-2902: First e2e test on feature nets in nightly pipeline (#1356)
# Description

This PR is the first part of moving e2e from PR pipeline to the nightly
pipeline. Essentially each e2e would be wrapped into
* start feature net
* run e2e test
* delete featurenet 

For now, they are run on short sessions. This also should change in the
future, as this is what we have nightly for (ie time for running normal
session tests).

Next tests would be added using `matrix`
(https://docs.github.com/en/actions/using-jobs/using-a-matrix-for-your-jobs)
in further PRs.

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)