A0-3053: Add logging to e2e button tests (#1366)
# Description

I made the tests run on a single node (I think some of the issues,
notably the timeouts were related to this) and added some logging around
the spots I think can still be flaky to get more info later. This might
not solve everything, but I think it's an improvement.

## Type of change

- Bug fix (non-breaking change which fixes an issue)