A0-2735: Change featurenets to use kustomize (#1378)
# Description

#### Changes featurenets to be created using kustomize instead of helm
charts
This is done to unify things a little bit and not copy
aleph-node-validator manifests but use the already-existing kustomize
base (with an overlay on top of it)

#### Fixes broken featurenets
Among few things the biggest bug that was introduced was appending Run
ID to the name of featurenet which caused plenty of infrastructure
problems (no sense to list them all here).

Appending Run ID is now optional for featurenet actions

#### Introduces `replicas` input for updatenets
Allows specific number of statefulset's replicas (meaning how many aleph
nodes are meant to be started). It's added both to `create` and
`update`, hence it is possible to scale up and down existing
featurenets.

## Type of change
It is an important fix.  Featurenets are broken anyway.