A0-3364: cache network details of validators (#1449)
# Description

Adding structures that will be later helpful to construct RPC call to
get network details of validators. This is a part 2/3 of the whole task,
see #1447). This only pulls and caches network data that was previously
observed by a node, without modifying or adding any new information flow
between nodes in the network.

Follow up to #1448 

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)

# Checklist:

<!-- delete when not applicable to your PR -->

- I have added tests
- I have created new documentation