A0-3390: Prepare for proposal compatibility (#1460)
# Description

Split `data_io` into two parts in a way which will make changing the
type of proposals much simpler. This is mostly moving/copying files
(`legacy` is essentially a copy of `data_io`), but some small additional
refactors also happened to make this simpler and the change less
invasive.

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)

# Checklist:

- I have made corresponding changes to the existing documentation
- I have created new documentation