A0-3010: Metrics for reorgs (#1483)
# Description

Adding metrics for (arbitrary) reorgs, as per initial slo doc. As for
the slo we care about the external user experience and what's visible to
they, I've decided to implement it in a separate thread, that uses
notification stream from the backend, the same as RPC does. I've also
moved best and top_finalized block to that task for the same reason.
After releasing we will have to update grafana to show `aleph_reorgs`
histogram somehow.

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

# Checklist:

- I have added tests
- I have made corresponding changes to the existing documentation