A0-3518: Resurrect chain extension: 2/3 (#1504)
# Description

Yet another PR, but it is not the end!

This time, I'm bringing back backend testing. Since the chain extension
mechanics are entangled with runtime and pallet contract, we have to
provide mocked environment, which results in a lot of lovely code with
unstable features. They were introduced because all the main calls
(`store_key` and `verify`) are invoked as **static** methods of
`Runtime`. Therefore, it makes testing different behaviors highly
inconvenient (since we cannot change the mock behavior depending on any
variable). Being able to introduce enums into generics gives us the
possibility of concise defining scenarios.

To keep PRs smaller, I left charging weight and validating input for the
next time.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

# Checklist:

- I have added tests
- I have made neccessary updates to the Infrastructure
- I have made corresponding changes to the existing documentation
- I have created new documentation