A0-3550: Better constants for libp2p message sizes (#1509)
# Description

Sync messages can be bigger, so increased the constant for them. Also
moved some code around to lower the likelyhood of making this code in
the future.

I also found some old compatibility code while going through this, so I
removed in on the way, it's definitely not needed any more.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

# Checklist:

- I have made corresponding changes to the existing documentation
- I have created new documentation