A0-3554: remove pruning option for rocksdb (#1516)
# Description

Removes `rocksdb` as the db back-end when one wants to use database
pruning. Pruning when `rocksdb` is used works differently than when we
use `paritydb`, i.e. some pruning state is stored in-memory instead of
being db-backed. In our tests, rocksdb-pruning was pretty unstable -
sometimes it works, sometimes it allocates whole available memory and
crashes. In case of `paritydb` we haven't observed such problems. It
also renames `--experimental-pruning` cli option to `enable-pruning`.

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

# Checklist:

- I have made corresponding changes to the existing documentation