A0-3015: Finality rate metrics (#1521)
# Description

Introduce `FinalityRateMetrics` - they measure the number of own
finalized blocks and the number of own hopeless blocks (our blocks which
are forking the finalized chain).

### Implementation details:
The interface consists of `{ report_own_imported, report_finalized }`.
`report_own_imported` is called when own block is imported and is added
to internal storage of `FinalityRateMetrics` object. `report_finalized`
is called when **any** finalized block appears and:
* if it is own block, it should have been added to internal storage when
reporting import - we count it as **own finalized**
* all blocks which are on the same level as the given newly finalized
block are removed from storage, and if they are not the given block,
they are counted as **own hopeless**.

## Type of change
- New feature (non-breaking change which adds functionality)

# Checklist:

<!-- delete when not applicable to your PR -->

- I have added tests
- I have made neccessary updates to the Infrastructure
- I have made corresponding changes to the existing documentation
- I have created new documentation