A0-2800: Push aleph-e2e-client image to ECR (#1526)
# Description

This PR adds a feature to push `aleph-e2e-client` to ECR, to reuse this
image in other repos. This image is pushed on commit to `main` or
`release-*` branch, also in nightly. Apart from that main feature, the
following tech debt was removed:
* extracted building `aleph-e2e-code` to separate action as it was
duplicated among repos,
* cleaned up image build workflows - there's a dedicated sub-workflow
for each type of node image
* merged two `self-hosted` jobs into one: `build test node and runtime`
and `build aleph-e2e-client image`. (speed-up: 3 minutes)
* increased retry time in `aleph-e2e-client` so that `sleep 60` is not
required (speed-up: 30 seconds)
* incorporating fix
https://github.com/Cardinal-Cryptography/github-actions/commit/8974638159c0439b223f7aa704eb434ee6e6b6fd

In total there's a 3-4 minute speed-up but it won't be visible yet as
the bottleneck is `check excluded packages` which will be addressed in
the further PRs.

## Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)
- New feature (non-breaking change which adds functionality)

## Testing:

Run workflows that are affected by this change except button one, as
that would be soon deprecated.
* on push to main:
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/7114087761/job/19367408317
* nightly tests on fe:
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/7113915981
* FE create:
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/7113918462
* nightly logic e2e tests:
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/7113958874
* nightly normal session tests:
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/7113921889