A0-3024: Transaction metrics (#1531)
# Description

This PR adds the _"Distribution of times which transactions took to be
included in a block"_ part of transaction metrics. For the _"Number of
new transactions in pool"_ metrics we could use
`substrate_sub_txpool_validations_finished` metrics from substrate.

For the former, I create a cache similar to the one we have for blocks.
By "to be included in a block" I assumed we meant first inclusion in any
valid block we've imported (and only first inclusion matters - see
tests), without waiting for its finalization. Also, for convenience, we
count only transactions that were ready right away when submitting (e.g.
with current nonce), no matter of whether they were spread through
network or submitted just locally.

The transaction pool provides only notifications about transactions
imported to the pool (validated), so checking if a transaction was
totally dropped from the pool or included in a block has to be done in
other way. My solution is as follows: when importing a block, we go
through all transactions that were included in that block, and report
them. ~From time to time, we also peek least recently used transactions
for the cache and remove them if they are no longer in the pool.~

This creates some race conditions, but since they shouldn't be very
frequent and distributed quite randomly over transactions, I think we
can ignore them. As an evidence, I've run a script that submits ~65
`vest` transactions per second to network created with
`scripts/run_nodes.sh`, with random submission time in the first second.
The graphs below show that:
* we miss ~17% transactions - those are because we've got (or are
handling) block import notification faster than the respective
transaction in pool notification. This also shows that we might be
underestimating the actual time from submitting transaction to block
inclusion by roughly 17% of block time = 170ms.
* rpc node has slightly "worse" (and closer to actual!) distribution,
because it gets a transaction faster and the block slower in
expectation.

Validations finished:

![validations_finished](https://github.com/Cardinal-Cryptography/aleph-node/assets/31205678/5cf567d1-f3c5-4d6f-9d76-e353334cf116)
Number of transactions reported:

![time_to_block_inclusion_number_of_observations](https://github.com/Cardinal-Cryptography/aleph-node/assets/31205678/f73cfb46-404a-4c7a-9aae-552cde0c8a8d)
Average time:

![average_time_to_block_inclusion](https://github.com/Cardinal-Cryptography/aleph-node/assets/31205678/12642b30-44b6-4346-a240-6b87fc7a3df2)


There is quite some mess with `chain_state_metrics` and `all_block`
metrics, and I think I can clean this up by moving the former to the
latter totally, but I'd rather do this in a separate task. I wonder what
are your thoughts about it.




## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)

# Checklist:

<!-- delete when not applicable to your PR -->

- I have added tests
- I have created new documentation