A0-3900: Sync handshake logic (#1592)
# Description

Add logic for handshake and peer slot constraints verification. The
verification part checks the genesis hash contained in the handshake,
and checks if we can connect the peer according to the specified
constraints. Added logic for peer sets management and counting the peers
of a particular role and direction (full inbound, full outbound, light).

The logic of accepting peers depending on their role and available spots
has been rewritten to be more readable and should preserve the previous
logic of `SyncingEngine`. The peer slot rules are:
* reserved nodes (those who have been specified during construction of
`Handler` as `reserved_nodes`) can freely connect and disconnect with no
constraints.
* there are separate bounds on the number of **inbound** and
**outbound** full nodes.
* there is a bound on the **light** nodes.
* The above bounds **ignore** the reserved nodes.

Removed the `mod test` because we can't construct the `Handler` now
without the `FullNetworkConfiguration`, and it's best to wait for the
substrate update before using it.

## Type of change
- New feature (non-breaking change which adds functionality)

# Checklist:
- I have made corresponding changes to the existing documentation